refactor: move menu classes to graphics package and add hardware acceleration toggle
