+++ src/main/java/de/athalion/game/twodgame/graphics/menu/MenuPage.java
@@ -1,0 +1,14 @@
+package de.athalion.game.twodgame.graphics.menu;
+
+import de.athalion.game.twodgame.input.KeyState;
+
+import java.awt.*;
+import java.awt.event.KeyEvent;
+
+public interface MenuPage {
+
+    void draw(Graphics2D g2);
+
+    MenuPage acceptInput(KeyState keyState, KeyEvent keyEvent);
+
+}
