--- src/main/java/de/athalion/game/twodgame/menu/MenuPage.java
@@ -1,14 +1,0 @@
-package de.athalion.game.twodgame.menu;
-
-import de.athalion.game.twodgame.input.KeyState;
-
-import java.awt.*;
-import java.awt.event.KeyEvent;
-
-public interface MenuPage {
-
-    void draw(Graphics2D g2);
-
-    MenuPage acceptInput(KeyState keyState, KeyEvent keyEvent);
-
-}
